:art: Remove dynamic imports, add README.
