+++ CHANGELOG.md
@@ -1,0 +1,5 @@
+# tiddlymark
+
+## 0.4.0 (10/26/2020)
+
+- [fs-plus]: refactor to avoid dynamic imports, add README
