+++ plugins/fs-plus/webpack.config.js
@@ -3,8 +3,9 @@ const { resolve } = require('path');
 module.exports = [
   {
     target: 'node',
+    node: 'empty',
     entry: {
-      'fs-plus/adapter': resolve(__dirname, `./src/adapter.ts`)
+      'fs-plus/adapter': resolve(__dirname, `./src/index.ts`)
     }
   }
 ];
